.kimi/skills/investment-skill/pyproject.toml
[project]
name = "investment-agent"
version = "1.0.0"
description = "Intelligent investment analysis system for Obsidian vault integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "Private"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Data Collection
    "akshare>=1.11.0",
    "yfinance>=0.2.28",
    "playwright>=1.40.0",
    "requests>=2.31.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Configuration & Templating
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    
    # Utilities
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
investment-agent = "__main__:main"
fund-monitor = "skills.fund_monitor:main"
macro-liquidity = "skills.macro_liquidity:main"
daily-summary = "skills.daily_summary:main"
semi-tracker = "skills.semiconductor_tracker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [".", "skills", "collectors", "utils"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
